flask/requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-texttospeech==0.3.0
googleapis-common-protos==1.5.6
grpcio==1.18.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
